backend/requirements.txt
Django>=4.2,<5
djangorestframework>=3.14,<4
django-cors-headers>=4.3,<5
psycopg[binary]>=3.1,<4
python-dotenv>=1.0,<2
django-environ>=0.11,<1
requests>=2.31,<3
Pillow>=10.0,<11
